+++ src/com/company/devices/Phone.java
@@ -4,7 +4,7 @@ public class Phone extends Device {
     Double screenSize;
     String os;
 
-    public Phone(String model, String producer, Integer value, Integer yearOfProduction, Double screenSize, String os) {
+    public Phone(String model, String producer, Double value, Integer yearOfProduction, Double screenSize, String os) {
         super(model, producer, value, yearOfProduction);
         this.screenSize = screenSize;
         this.os = os;
